fix: step 관리에 대한 중복된 로직 수정
