+++ src/features/event-manage/ui/evnetCreation/EventFunnel.tsx
@@ -9,6 +9,7 @@ import EventTypePage from '../../../../pages/event-manage/ui/EventTypePage';
 import EventTagPage from '../../../../pages/event-manage/ui/EventTagPage';
 import EventOrganizerInfo from '../../../../pages/event-manage/ui/EventOrganizerInfo';
 import EventRegisterLayout from '../../../../shared/ui/backgrounds/EventRegisterLayout';
+import { useNavigate } from 'react-router-dom';
 
 export interface EventFunnelInterface {
   steps: string[];
@@ -16,10 +17,12 @@ export interface EventFunnelInterface {
   onPrev: (prevStep: string) => void;
   Funnel: React.ComponentType<FunnelProps>;
   Step: React.ComponentType<StepProps>;
+  currentStep: number;
 }
 
-const EventFunnel = ({ steps, onNext, onPrev, Funnel, Step }: EventFunnelInterface) => {
+const EventFunnel = ({ onNext, onPrev, Funnel, Step, currentStep }: EventFunnelInterface) => {
   const { data } = useFunnelStore();
+  const navigate = useNavigate();
 
   const handleNext = (nextStep: string) => {
     console.log('이벤트 주최 데이터:', data);
@@ -31,8 +34,8 @@ const EventFunnel = ({ steps, onNext, onPrev, Funnel, Step }: EventFunnelInterfa
       <Step name="HostSelection">
         <EventRegisterLayout
           title="이벤트를 호스팅할 채널을 선택해주세요"
-          onNext={() => handleNext(steps[0])}
-          // onPrev={() => {}}
+          onNext={() => handleNext(String(currentStep + 1))}
+          onPrev={() => navigate(-1)}
         >
           <HostSelectionPage />
         </EventRegisterLayout>
@@ -40,8 +43,8 @@ const EventFunnel = ({ steps, onNext, onPrev, Funnel, Step }: EventFunnelInterfa
       <Step name="EventTitle">
         <EventRegisterLayout
           title="이벤트 제목을 입력해주세요"
-          onNext={() => handleNext(steps[1])}
-          onPrev={() => onPrev(steps[0])}
+          onNext={() => handleNext(String(currentStep + 1))}
+          onPrev={() => onPrev(String(currentStep - 1))}
         >
           <EventTitlePage />
         </EventRegisterLayout>
@@ -49,8 +52,8 @@ const EventFunnel = ({ steps, onNext, onPrev, Funnel, Step }: EventFunnelInterfa
       <Step name="EventPeriod">
         <EventRegisterLayout
           title="이벤트 기간을 입력해주세요"
-          onNext={() => handleNext(steps[2])}
-          onPrev={() => onPrev(steps[1])}
+          onNext={() => handleNext(String(currentStep + 1))}
+          onPrev={() => onPrev(String(currentStep - 1))}
         >
           <EventPeriodPage />
         </EventRegisterLayout>
@@ -58,8 +61,8 @@ const EventFunnel = ({ steps, onNext, onPrev, Funnel, Step }: EventFunnelInterfa
       <Step name="EventOrganizerInfo">
         <EventRegisterLayout
           title="이벤트 주최자 정보를 입력해주세요"
-          onNext={() => handleNext(steps[3])}
-          onPrev={() => onPrev(steps[2])}
+          onNext={() => handleNext(String(currentStep + 1))}
+          onPrev={() => onPrev(String(currentStep - 1))}
         >
           <EventOrganizerInfo />
         </EventRegisterLayout>
@@ -67,8 +70,8 @@ const EventFunnel = ({ steps, onNext, onPrev, Funnel, Step }: EventFunnelInterfa
       <Step name="EventInfo">
         <EventRegisterLayout
           title="이벤트 정보를 입력해주세요"
-          onNext={() => handleNext(steps[4])}
-          onPrev={() => onPrev(steps[3])}
+          onNext={() => handleNext(String(currentStep + 1))}
+          onPrev={() => onPrev(String(currentStep - 1))}
         >
           <EventInfoPage />
         </EventRegisterLayout>
@@ -76,14 +79,17 @@ const EventFunnel = ({ steps, onNext, onPrev, Funnel, Step }: EventFunnelInterfa
       <Step name="EventType">
         <EventRegisterLayout
           title="이벤트 진행방식을 선택해주세요"
-          onNext={() => handleNext(steps[5])}
-          onPrev={() => onPrev(steps[4])}
+          onNext={() => handleNext(String(currentStep + 1))}
+          onPrev={() => onPrev(String(currentStep - 1))}
         >
           <EventTypePage />
         </EventRegisterLayout>
       </Step>
       <Step name="EventTag">
-        <EventRegisterLayout onNext={() => handleNext(steps[6])} onPrev={() => onPrev(steps[5])}>
+        <EventRegisterLayout
+          onNext={() => handleNext(String(currentStep + 1))}
+          onPrev={() => onPrev(String(currentStep - 1))}
+        >
           <EventTagPage />
         </EventRegisterLayout>
       </Step>
